+++ angadi/angadi/src/main/java/com/angadi/service/WalletService.java
@@ -8,9 +8,9 @@ public interface WalletService {
 
     public Wallet addWallet(Wallet wallet, String email) throws CustomerException;
 
-    public Wallet addBalanceToWallet(Double amount, String email) throws CustomerException, WalletException;
+    public Wallet addBalanceToWallet(Integer amount, String email) throws CustomerException, WalletException;
 
-    public Double showBalance(Integer walletId, String email) throws WalletException, CustomerException;
+    public Integer showBalance(Integer walletId, String email) throws WalletException, CustomerException;
 
-    public Wallet transferAmount(Integer walletId, String description, Double amount, String email) throws WalletException, CustomerException;
+    public Wallet transferAmount(Integer walletId, String description, Integer amount, String email) throws WalletException, CustomerException;
 }
